glusterfsd creates pidfile iff -p option is given

+++ glusterfsd/src/glusterfsd.c
@@ -62,7 +62,7 @@ static int parse_opts (int32_t key, char *arg, struct argp_state *_state);
 static struct argp argp = { options, parse_opts, argp_doc, doc };
 
 static char *specfile = CONFDIR "/glusterfs-server.vol";
-static char *pidfile = DATADIR "/run/glusterfsd.pid";
+static char *pidfile = NULL;
 static xlator_t *xlator_tree_node = NULL;
 
 static xlator_t *
@@ -205,7 +205,10 @@ void
 glusterfsd_cleanup_and_exit (int signum)
 {
   gf_log ("glusterfsd", GF_LOG_WARNING, "shutting down server");
-  unlink (pidfile);
+
+  if (pidfile)
+    unlink (pidfile);
+
   exit (0);
 }
 
@@ -227,7 +230,8 @@ main (int32_t argc, char *argv[])
 
   argp_parse (&argp, argc, argv, 0, 0, &ctx);
 
-  pidfd = pidfile_lock (pidfile);
+  if (pidfile)
+    pidfd = pidfile_lock (pidfile);
 
   if (gf_log_init (ctx.logfile) < 0){
     return 1;
@@ -282,7 +286,9 @@ main (int32_t argc, char *argv[])
       memset (argv[i], ' ', strlen (argv[i]));
     sprintf (argv[0], "[glusterfsd]");
     daemon (0, 0);
-    pidfile_update (pidfd);
+
+    if (pidfile)
+      pidfile_update (pidfd);
   } 
 
   gf_timer_registry_init (&ctx);
@@ -309,6 +315,8 @@ main (int32_t argc, char *argv[])
 
   while (!poll_iteration (&ctx));
 
-  close (pidfd);
+  if (pidfile)
+    close (pidfd);
+
   return 0;
 }
